HDFS-17117. Modify patch based on comments

+++ hadoop-hdfs-project/hadoop-hdfs-rbf/src/main/java/org/apache/hadoop/hdfs/server/federation/metrics/NamenodeBeanMetrics.java
@@ -885,8 +885,8 @@ public class NamenodeBeanMetrics
   }
 
   @Override
-  public double getReconstructionQueuesInitProgress() {
-    return 0.0;
+  public float getReconstructionQueuesInitProgress() {
+    return 0;
   }
 
   private Router getRouter() throws IOException {
