+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/namenode/FSNamesystem.java
@@ -4893,7 +4893,7 @@ public class FSNamesystem implements Namesystem, FSNamesystemMBean,
    */
   @Override // FSNamesystemMBean
   @Metric
-  public double getReconstructionQueuesInitProgress() {
+  public float getReconstructionQueuesInitProgress() {
     return blockManager.getReconstructionQueuesInitProgress();
   }
